renamed access_type to access_types. Added a getAccessTypes which will return the latest accessType record from the table.
