+++ src/main/java/uk/gov/hmcts/reform/orgrolemapping/data/AccessTypesRepository.java
@@ -6,7 +6,12 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface AccessTypesRepository extends CrudRepository<AccessTypes, Long> {
 
-    AccessTypes findByAccessType(String json);
-
+    //retrieves the AccessTypes record by version.
     AccessTypes findByVersion(Long version);
+
+    // Retrieves the first AccessTypes record.
+    // 1. It first sorts the AccessTypes records in descending order by version.
+    // 2.Hence, we get the top version among all the AccessTypes
+    AccessTypes findFirstByOrderByVersionDesc();
+
 }
