+++ src/main/java/uk/gov/hmcts/reform/orgrolemapping/domain/service/AccessTypeService.java
@@ -5,20 +5,20 @@ import uk.gov.hmcts.reform.orgrolemapping.data.AccessTypes;
 import uk.gov.hmcts.reform.orgrolemapping.data.AccessTypesRepository;
 
 @Service
-public class PersistenceAccessTypeService {
+public class AccessTypeService {
+
 
     private final AccessTypesRepository accessTypesRepository;
 
-    public PersistenceAccessTypeService(AccessTypesRepository accessTypesRepository) {
+    public AccessTypeService(AccessTypesRepository accessTypesRepository) {
         this.accessTypesRepository = accessTypesRepository;
-    }
-
-
-    public AccessTypes getAccessType(String json) {
-        return accessTypesRepository.findByAccessType(json);
     }
 
     public AccessTypes getVersion(Long version) {
         return accessTypesRepository.findByVersion(version);
+    }
+
+    public AccessTypes getAccessTypes(){
+        return accessTypesRepository.findFirstByOrderByVersionDesc();
     }
 }
