+++ src/main/resources/db/migration/V20231218_47__GA-47_new_access_types_table.sql
@@ -1,5 +1,5 @@
 -- create table
--- PK/FK relationships in tables access_types
+-- PK/FK relationships in table access_types
 -- access_types pk column = version
 -- automatic when creating bigserial:
 --   access_types_version_seq (Sequence)
@@ -7,7 +7,7 @@
 
 CREATE TABLE access_types(
 	version bigserial not null,
-    access_type jsonb NOT NULL,
+    access_types jsonb NOT NULL,
 	CONSTRAINT pk_access_types PRIMARY KEY (version)
 );
